Extract runtime contract test harnesses
