+++ packages/runtime/src/kafka-ingress.internal.test.ts
@@ -1,17 +1,6 @@
 import { describe, expect, it } from "@effect/vitest";
 import { Consumer } from "@platformatic/kafka";
-import type { Message } from "@platformatic/kafka";
-import {
-  defineViewServerConfig,
-  kafka,
-  type RuntimeRegions,
-  type ViewServerConfig,
-  type ViewServerRuntimeError,
-} from "@effect-view-server/config";
-import {
-  makeKafkaSourceTopicsForConfig,
-  type KafkaResolvedSourceTopicDefinition,
-} from "@effect-view-server/config/internal";
+import { defineViewServerConfig, kafka } from "@effect-view-server/config";
 import {
   makeViewServerRuntimeCoreInternal as makeViewServerRuntimeCore,
   type ViewServerRuntimeCoreInternalClient,
@@ -29,10 +18,8 @@ import {
   Option,
   Queue,
   References,
-  Schema,
   Scope,
 } from "effect";
-import { makeViewServerKafkaHealthLedger as makeViewServerKafkaHealthLedgerBase } from "./kafka-health";
 import type { ViewServerKafkaHealthLedger } from "./kafka-health";
 import {
   assignedPartitionsForSourceTopic,
@@ -48,7 +35,6 @@ import {
   kafkaHeadersFromMessage,
   kafkaConsumerCloseError,
   kafkaConsumerStartError,
-  kafkaIngressErrorSourceTopic,
   kafkaMessageCommitError,
   kafkaMessageDecodeError,
   kafkaMessageProcessingError,
@@ -80,318 +66,36 @@ import type {
 } from "./kafka-ingress";
 import type { ResolvedViewServerKafkaRuntimeOptions } from "./runtime-options";
 import { resolveViewServerRuntimeOptions } from "./runtime-options";
-import type { ViewServerRuntimeTopicDefinitions } from "./runtime-types";
-
-class KafkaIngressTestError extends Schema.TaggedErrorClass<KafkaIngressTestError>()(
-  "KafkaIngressTestError",
-  {
-    message: Schema.String,
-  },
-) {}
-
-const Order = Schema.Struct({
-  id: Schema.String,
-  customerId: Schema.String,
-  price: Schema.Number,
-});
-
-const IncomingOrder = Schema.Struct({
-  customerId: Schema.String,
-  price: Schema.Number,
-});
-
-const PrecisePosition = Schema.Struct({
-  id: Schema.String,
-  accountId: Schema.String,
-  quantity: Schema.BigInt,
-  price: Schema.BigDecimal,
-});
-
-const IncomingPrecisePosition = Schema.Struct({
-  accountId: Schema.String,
-  quantity: Schema.BigInt,
-  price: Schema.BigDecimal,
-});
-
-const regions = {
-  cold: "localhost:9093",
-  local: " localhost:9092, ,localhost:9094 ",
-};
-const ordersSourceTopic = "orders-source";
-const paymentsSourceTopic = "payments-source";
-const unknownSourceTopic = "unknown-source";
-
-const viewServer = defineViewServerConfig({
-  kafka: regions,
-  topics: {
-    orders: {
-      schema: Order,
-      key: "id",
-      kafkaSource: kafka.source({
-        topic: ordersSourceTopic,
-        regions: ["local"],
-        value: kafka.json(IncomingOrder),
-        key: kafka.stringKey(),
-        rowKey: ({ key }) => key,
-        map: ({ value }) => ({
-          customerId: value.customerId,
-          price: value.price,
-        }),
-      }),
-    },
-    precisePositions: {
-      schema: PrecisePosition,
-      key: "id",
-    },
-  },
-});
-
-type Topics = typeof viewServer.topics;
-type KafkaMessageBytes = Buffer | null | undefined;
-type KafkaMessage = Message<KafkaMessageBytes, KafkaMessageBytes, Buffer, Buffer>;
-type CapturedLog = {
-  readonly cause: Cause.Cause<unknown>;
-  readonly message: unknown;
-};
-
-const makeCapturedLogs = () => {
-  const logs: Array<CapturedLog> = [];
-  const logger = Logger.make<unknown, void>((options) => {
-    logs.push({
-      cause: options.cause,
-      message: options.message,
-    });
-  });
-  return { logger, logs };
-};
-
-const nonStringTagCodecError: { readonly _tag: 123; readonly message: "non-string tag" } = {
-  _tag: 123,
-  message: "non-string tag",
-};
-const forgedMappingTagCodecError: {
-  readonly _tag: "KafkaMappingError";
-  readonly [key: symbol]: symbol;
-  readonly message: "forged mapping tag";
-} = {
-  _tag: "KafkaMappingError",
-  [Symbol.for("@effect-view-server/config/KafkaMappingError")]: Symbol.for(
-    "@effect-view-server/config/KafkaMappingError",
-  ),
-  message: "forged mapping tag",
-};
-
-const makeCommittedKafkaStart = (consumerGroupId: string) => ({
-  consume: {
-    consumerGroupId,
-    fallbackMode: "earliest" as const,
-    mode: "committed" as const,
-  },
-  startFrom: {
-    committedConsumerGroup: consumerGroupId,
-  },
-});
-
-const committedKafkaStart = (
-  consumerGroupId: string,
-): Pick<ResolvedViewServerKafkaRuntimeOptions<Topics>, "consume" | "startFrom"> =>
-  makeCommittedKafkaStart(consumerGroupId);
-
-const makeRuntimeTopicRecord = <
-  const CurrentTopics extends ViewServerRuntimeTopicDefinitions,
-  const CurrentRegions extends RuntimeRegions,
->(
-  config: ViewServerConfig<CurrentTopics, CurrentRegions>,
-): Record<string, KafkaResolvedSourceTopicDefinition<CurrentTopics, CurrentRegions>> => {
-  const topics: Record<
-    string,
-    KafkaResolvedSourceTopicDefinition<CurrentTopics, CurrentRegions>
-  > = Object.create(null);
-  for (const topic of makeKafkaSourceTopicsForConfig<CurrentTopics, CurrentRegions>(config)) {
-    topics[topic.topic] = topic;
-  }
-  return topics;
-};
-
-const kafkaOptionsForConfig = <
-  const CurrentTopics extends ViewServerRuntimeTopicDefinitions,
-  const CurrentRegions extends RuntimeRegions,
->(
-  config: ViewServerConfig<CurrentTopics, CurrentRegions>,
-  consumerGroupId: string,
-  runtimeRegions: Record<string, string> = regions,
-): ResolvedViewServerKafkaRuntimeOptions<CurrentTopics, CurrentRegions> => ({
-  consumerGroupId,
-  ...makeCommittedKafkaStart(consumerGroupId),
-  regions: runtimeRegions,
-  topics: makeRuntimeTopicRecord(config),
-});
-
-const kafkaOptions: ResolvedViewServerKafkaRuntimeOptions<Topics> = {
-  consumerGroupId: "view-server-test",
-  ...committedKafkaStart("view-server-test"),
+import {
+  causeReasonSummary,
+  committedKafkaStart,
+  decodeFailureThenSuccessKafkaStream,
+  failingClient,
+  failingKafkaStream,
+  forgedMappingTagCodecError,
+  IncomingOrder,
+  IncomingPrecisePosition,
+  KafkaIngressTestError,
+  kafkaIngressErrorSourceTopicOrNull,
+  kafkaIngressErrorSummary,
+  kafkaMessage,
+  kafkaOptions,
+  kafkaOptionsForConfig,
+  makeCapturedLogs,
+  makeViewServerKafkaHealthLedger,
+  nonStringTagCodecError,
+  nullRecord,
+  Order,
+  ordersSourceTopic,
+  paymentsSourceTopic,
+  PrecisePosition,
   regions,
-  topics: makeRuntimeTopicRecord(viewServer),
-};
-
-type KafkaHealthLedgerInput<LedgerTopics extends ViewServerRuntimeTopicDefinitions> = Parameters<
-  typeof makeViewServerKafkaHealthLedgerBase<LedgerTopics>
->[0];
-
-const makeViewServerKafkaHealthLedger = <
-  const LedgerTopics extends ViewServerRuntimeTopicDefinitions,
->(
-  input: Omit<KafkaHealthLedgerInput<LedgerTopics>, "startFrom"> &
-    Partial<Pick<KafkaHealthLedgerInput<LedgerTopics>, "startFrom">>,
-): ViewServerKafkaHealthLedger<LedgerTopics> =>
-  makeViewServerKafkaHealthLedgerBase<LedgerTopics>({
-    startFrom: kafkaOptions.consume,
-    ...input,
-  });
-
-const nullRecord = <Value>(entries: Record<string, Value>): Record<string, Value> => {
-  const record: Record<string, Value> = Object.create(null);
-  return Object.assign(record, entries);
-};
-
-const causeReasonSummary = (
-  cause: unknown,
-): ReadonlyArray<{
-  readonly tag: string;
-  readonly message: string | null;
-}> =>
-  Cause.isCause(cause)
-    ? cause.reasons.map((reason) => ({
-        tag: reason._tag,
-        message:
-          reason._tag === "Fail"
-            ? messageFromUnknown(reason.error)
-            : reason._tag === "Die"
-              ? messageFromUnknown(reason.defect)
-              : null,
-      }))
-    : [];
-
-const failedKafkaStreamQueueEvent = (
-  event: KafkaStreamQueueEvent,
-): Effect.Effect<Extract<KafkaStreamQueueEvent, { readonly _tag: "Failed" }>, Error> =>
-  Effect.succeed(event).pipe(
-    Effect.filterOrFail(
-      (value): value is Extract<KafkaStreamQueueEvent, { readonly _tag: "Failed" }> =>
-        value._tag === "Failed",
-      () => new KafkaIngressTestError({ message: "Expected Kafka stream queue failure event." }),
-    ),
-  );
-
-const kafkaIngressErrorSummary = (error: unknown) =>
-  error instanceof ViewServerKafkaIngressError
-    ? {
-        cause: causeReasonSummary(error.cause),
-        message: error.message,
-        region: error.region,
-        sourceTopic: error.sourceTopic,
-      }
-    : null;
-
-const kafkaIngressErrorSourceTopicOrNull = (error: unknown): string | null =>
-  Option.getOrNull(kafkaIngressErrorSourceTopic(error));
-
-const kafkaMessage = (input: {
-  readonly topic: string;
-  readonly key?: string | null;
-  readonly value?: string | null;
-  readonly headers?: ReadonlyMap<Buffer, Buffer>;
-  readonly offset?: bigint;
-  readonly onCommit?: () => void;
-  readonly commitFailure?: Error;
-}): KafkaMessage => {
-  const headers = new Map(input.headers ?? []);
-  const key = input.key === undefined || input.key === null ? input.key : Buffer.from(input.key);
-  const value =
-    input.value === undefined || input.value === null ? input.value : Buffer.from(input.value);
-  const offset = input.offset ?? 0n;
-  return {
-    key,
-    value,
-    headers,
-    topic: input.topic,
-    partition: 0,
-    timestamp: 1_234n,
-    offset,
-    metadata: {},
-    commit: () => {
-      if (input.commitFailure !== undefined) {
-        return Promise.reject(input.commitFailure);
-      }
-      input.onCommit?.();
-      return undefined;
-    },
-    toJSON: () => ({
-      key,
-      value,
-      headers: Array.from(headers.entries()),
-      topic: input.topic,
-      partition: 0,
-      timestamp: "1234",
-      offset: String(offset),
-      metadata: {},
-    }),
-  };
-};
-
-const runtimeUnavailable: ViewServerRuntimeError = {
-  _tag: "ViewServerRuntimeError",
-  code: "RuntimeUnavailable",
-  message: "publish failed",
-};
-
-const failingClient: ViewServerRuntimeCoreInternalClient<Topics> = {
-  delete: () => Effect.fail(runtimeUnavailable),
-  health: () => Effect.fail(runtimeUnavailable),
-  patch: () => Effect.fail(runtimeUnavailable),
-  publish: () => Effect.fail(runtimeUnavailable),
-  publishMany: () => Effect.fail(runtimeUnavailable),
-  publishManyDecodedRows: () => Effect.fail(runtimeUnavailable),
-  publishManyDecodedRowsWithStorageKeys: () => Effect.fail(runtimeUnavailable),
-  publishManyWithStorageKeys: () => Effect.fail(runtimeUnavailable),
-  reset: () => Effect.fail(runtimeUnavailable),
-  snapshot: () => Effect.fail(runtimeUnavailable),
-};
-
-async function* failingKafkaStream(): AsyncIterable<KafkaMessage> {
-  yield kafkaMessage({
-    topic: ordersSourceTopic,
-    key: "order-stream-1",
-    value: JSON.stringify({
-      customerId: "customer-stream-1",
-      price: 30,
-    }),
-    offset: 4n,
-  });
-  throw new Error("stream-down");
-}
-
-async function* decodeFailureThenSuccessKafkaStream(
-  onCommit: () => void,
-): AsyncIterable<KafkaMessage> {
-  yield kafkaMessage({
-    topic: ordersSourceTopic,
-    key: "bad-json",
-    value: "{",
-    offset: 1n,
-    onCommit,
-  });
-  yield kafkaMessage({
-    topic: ordersSourceTopic,
-    key: "order-after-failure",
-    value: JSON.stringify({
-      customerId: "customer-after-failure",
-      price: 40,
-    }),
-    offset: 2n,
-    onCommit,
-  });
-}
+  runtimeUnavailable,
+  type KafkaMessage,
+  type Topics,
+  unknownSourceTopic,
+  viewServer,
+} from "../test-harness/kafka-ingress";
 
 describe("@effect-view-server/runtime Kafka ingress internals", () => {
   it("normalizes Kafka helper values", () => {
@@ -3883,7 +3587,16 @@ describe("@effect-view-server/runtime Kafka ingress internals", () => {
         queue,
         Cause.fromReasons([Cause.makeFailReason(rawFailure), Cause.makeInterruptReason()]),
       );
-      const event = yield* Queue.take(queue).pipe(Effect.andThen(failedKafkaStreamQueueEvent));
+      const event = yield* Queue.take(queue).pipe(
+        Effect.filterOrFail(
+          (value): value is Extract<KafkaStreamQueueEvent, { readonly _tag: "Failed" }> =>
+            value._tag === "Failed",
+          () =>
+            new KafkaIngressTestError({
+              message: "Expected Kafka stream queue failure event.",
+            }),
+        ),
+      );
 
       expect({
         cause: causeReasonSummary(event.cause),
@@ -3929,7 +3642,16 @@ describe("@effect-view-server/runtime Kafka ingress internals", () => {
       ];
 
       yield* offerKafkaStreamProducerFailure("local", queue, Cause.fromReasons(failureReasons));
-      const event = yield* Queue.take(queue).pipe(Effect.andThen(failedKafkaStreamQueueEvent));
+      const event = yield* Queue.take(queue).pipe(
+        Effect.filterOrFail(
+          (value): value is Extract<KafkaStreamQueueEvent, { readonly _tag: "Failed" }> =>
+            value._tag === "Failed",
+          () =>
+            new KafkaIngressTestError({
+              message: "Expected Kafka stream queue failure event.",
+            }),
+        ),
+      );
 
       expect({
         cause: causeReasonSummary(event.cause),
